refactor: add cNBT library
