+++ libraries/cNBT/include/cNBT/buffer.h
