+++ libraries/cNBT/include/cNBT/list.h
