+++ libraries/cNBT/source/nbt_treeops.c
@@ -7,7 +7,7 @@
  * it, you can buy us a beer in return.
  * -----------------------------------------------------------------------------
  */
-#include "nbt.h"
+#include <cNBT/nbt.h>
 
 #include <assert.h>
 #include <errno.h>
