+++ libraries/cNBT/source/nbt_util.c
@@ -7,7 +7,7 @@
  * it, you can buy us a beer in return.
  * -----------------------------------------------------------------------------
  */
-#include "nbt.h"
+#include <cNBT/nbt.h>
 
 #include <string.h>
